+++ src/Api/Controllers/AccountsController.cs
@@ -626,6 +626,7 @@ namespace Bit.Api.Controllers
             };
         }
 
+        [Obsolete("2022-04-01 Use separate Billing History/Payment APIs, left for backwards compatability with older clients")]
         [HttpGet("billing")]
         [SelfHosted(NotSelfHostedOnly = true)]
         public async Task<BillingResponseModel> GetBilling()
@@ -639,35 +640,7 @@ namespace Bit.Api.Controllers
             var billingInfo = await _paymentService.GetBillingAsync(user);
             return new BillingResponseModel(billingInfo);
         }
-
-        [HttpGet("billing-history")]
-        [SelfHosted(NotSelfHostedOnly = true)]
-        public async Task<BillingHistoryResponseModel> GetBillingHistory()
-        {
-            var user = await _userService.GetUserByPrincipalAsync(User);
-            if (user == null)
-            {
-                throw new UnauthorizedAccessException();
-            }
-
-            var billingInfo = await _paymentService.GetBillingHistoryAsync(user);
-            return new BillingHistoryResponseModel(billingInfo);
-        }
-
-        [HttpGet("billing-payment")]
-        [SelfHosted(NotSelfHostedOnly = true)]
-        public async Task<BillingPaymentResponseModel> GetPaymentMethod()
-        {
-            var user = await _userService.GetUserByPrincipalAsync(User);
-            if (user == null)
-            {
-                throw new UnauthorizedAccessException();
-            }
-
-            var billingInfo = await _paymentService.GetBillingBalanceAndSourceAsync(user);
-            return new BillingPaymentResponseModel(billingInfo);
-        }
-
+        
         [HttpGet("subscription")]
         public async Task<SubscriptionResponseModel> GetSubscription()
         {
